# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/connect:
    post:
      operationId: connectDerivAccount
      tags: [auth]
      summary: Connect a Deriv account via API token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DerivTokenInput"
      responses:
        "200":
          description: Connected account info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DerivAccount"
        "400":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/account:
    get:
      operationId: getAccount
      tags: [auth]
      summary: Get connected account details
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DerivAccount"
        "404":
          description: No account connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /markets:
    get:
      operationId: getMarkets
      tags: [markets]
      summary: Get all ranked markets with AI quality scores
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [all, synthetic, forex, commodities, derived]
          description: Filter by market category
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Ranked market list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RankedMarket"
  /markets/{symbol}:
    get:
      operationId: getMarketDetail
      tags: [markets]
      summary: Get detailed analysis for a specific market
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Market detail with full agent scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketDetail"
        "404":
          description: Market not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: Get trade journal with filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, won, lost, open]
        - name: market
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trade list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"

    post:
      operationId: executeTrade
      tags: [trades]
      summary: Execute a manual or AI-recommended trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Trade executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "400":
          description: Trade rejected by risk controls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a single trade by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    DerivTokenInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
    DerivAccount:
      type: object
      required: [id, loginId, currency, balance, isVirtual]
      properties:
        id:
          type: integer
        loginId:
          type: string
        currency:
          type: string
        balance:
          type: number
        isVirtual:
          type: boolean
        email:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        connectedAt:
          type: string
    RankedMarket:
      type: object
      required: [symbol, displayName, category, qualityScore, confidenceScore, riskScore, trend, volatility, rank]
      properties:
        symbol:
          type: string
        displayName:
          type: string
        category:
          type: string
          enum: [synthetic, forex, commodities, derived]
        qualityScore:
          type: number
          description: Composite market quality score 0-100
        confidenceScore:
          type: number
          description: AI confidence that a good trade exists 0-100
        riskScore:
          type: number
          description: Risk level 0-100 (lower is safer)
        trend:
          type: string
          enum: [strong_up, up, sideways, down, strong_down]
        volatility:
          type: string
          enum: [low, medium, high, extreme]
        rank:
          type: integer
        recommendedContractType:
          type: ["string", "null"]
        lastPrice:
          type: ["number", "null"]
        priceChange24h:
          type: ["number", "null"]
    MarketDetail:
      type: object
      required: [symbol, displayName, category, qualityScore, agentScores, recommendation, priceHistory]
      properties:
        symbol:
          type: string
        displayName:
          type: string
        category:
          type: string
        qualityScore:
          type: number
        agentScores:
          $ref: "#/components/schemas/AgentScores"
        recommendation:
          $ref: "#/components/schemas/AiRecommendation"
        priceHistory:
          type: array
          items:
            $ref: "#/components/schemas/PricePoint"
        lastUpdated:
          type: string
    AgentScores:
      type: object
      required: [marketScanner, trendAnalysis, volatilityAnalysis, patternRecognition, riskManagement, capitalPreservation, tradeExecution, selfLearning]
      properties:
        marketScanner:
          $ref: "#/components/schemas/AgentScore"
        trendAnalysis:
          $ref: "#/components/schemas/AgentScore"
        volatilityAnalysis:
          $ref: "#/components/schemas/AgentScore"
        patternRecognition:
          $ref: "#/components/schemas/AgentScore"
        riskManagement:
          $ref: "#/components/schemas/AgentScore"
        capitalPreservation:
          $ref: "#/components/schemas/AgentScore"
        tradeExecution:
          $ref: "#/components/schemas/AgentScore"
        selfLearning:
          $ref: "#/components/schemas/AgentScore"
    AgentScore:
      type: object
      required: [score, weight, signal, reasoning]
      properties:
        score:
          type: number
        weight:
          type: number
        signal:
          type: string
          enum: [strong_buy, buy, neutral, sell, strong_sell]
        reasoning:
          type: string
    PricePoint:
      type: object
      required: [timestamp, price]
      properties:
        timestamp:
          type: string
        price:
          type: number
...
    Trade:
      type: object
      required: [id, symbol, contractType, stake, direction, status, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        contractType:
          type: string
        stake:
          type: number
        direction:
          type: string
          enum: [up, down]
        status:
          type: string
          enum: [open, won, lost, cancelled]
        payout:
          type: ["number", "null"]
        profit:
          type: ["number", "null"]
        entryPrice:
          type: ["number", "null"]
        exitPrice:
          type: ["number", "null"]
        aiConfidence:
          type: ["number", "null"]
        aiRiskScore:
          type: ["number", "null"]
        isAutonomous:
          type: boolean
        agentReasoning:
          type: ["string", "null"]
        createdAt:
          type: string
        closedAt:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
    TradeInput:
      type: object
      required: [symbol, contractType, stake, direction]
      properties:
        symbol:
          type: string
        contractType:
          type: string
        stake:
          type: number
        direction:
          type: string
          enum: [up, down]
        isAutonomous:
          type: boolean
        duration:
          type: integer
        durationUnit:
          type: string
          enum: [t, s, m, h, d]
...
    AiRecommendation:
      type: object
      required: [symbol, contractType, direction, stake, confidence, riskScore, profitability, agentScores, shouldTrade, reasoning, warnings]
      properties:
        symbol:
          type: string
        contractType:
          type: string
        direction:
          type: string
          enum: [up, down]
        stake:
          type: number
        confidence:
          type: number
        riskScore:
          type: number
        profitability:
          type: number
        agentScores:
          $ref: "#/components/schemas/AgentScores"
        shouldTrade:
          type: boolean
        reasoning:
          type: string
        warnings:
          type: array
          items:
            type: string
        generatedAt:
          type: string
...