# lib/api-zod/src/generated/types/aiRecommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentScores } from './agentScores';
import type { AiRecommendationDirection } from './aiRecommendationDirection';

export interface AiRecommendation {
  symbol: string;
  contractType: string;
  direction: AiRecommendationDirection;
  stake: number;
  confidence: number;
  riskScore: number;
  profitability: number;
  agentScores: AgentScores;
  shouldTrade: boolean;
  reasoning: string;
  warnings: string[];
  generatedAt?: string;
}