# lib/api-zod/src/generated/types/tradeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TradeInputDirection } from './tradeInputDirection';
import type { TradeInputDurationUnit } from './tradeInputDurationUnit';

export interface TradeInput {
  symbol: string;
  contractType: string;
  stake: number;
  direction: TradeInputDirection;
  isAutonomous?: boolean;
  duration?: number;
  durationUnit?: TradeInputDurationUnit;
}